First commit of the blobs update - HTC INC
This commit is the first commit of the blobs update for 2.3.4 for
the HTC incredible. Only includes the updated blobs, all references
are final, but the directory structure is final minus the BACKUP
directory in case I mess up.

TODO: This commit should be sqaushed with the next commits so that
the blobs are updated in one commit.

Remove old blobs location